+++ CHANGELOG.md
@@ -4,6 +4,12 @@ This file documents all notable changes to this project. The format of this file
 is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this
 project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Added saving column descriptions for multidimensional clustering into db.
+
 ## [0.6.2] - 2019-10-11
 
 ### Added